crawler/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
rich==13.7.1
